[xrhome] Switch to local source of truth for image targets
