+++ reality/cloud/xrhome/src/client/apps/image-targets/image-helpers.ts
@@ -1,6 +1,6 @@
 import {MINIMUM_SHORT_LENGTH, MINIMUM_LONG_LENGTH} from '../../../shared/xrengine-config'
 import type {CropAreaPixels} from '../../common/image-cropper'
-import type {CanvasPool, Resource, ResourcePool} from '../../common/resource-pool'
+import type {CanvasPool} from '../../common/resource-pool'
 import {computePixelPointsFromRadius, getUnconifiedData} from './conical/unconify'
 
 type ImageInfo = {url: string, data: ImageData}
@@ -48,10 +48,7 @@ const getDefaultBottomRadius = (width, height, rTop): number => {
   return Math.min(rBottom, rTop - minUnconifiedHeight)
 }
 
-const loadImage = (
-  url: string,
-  imgPool: ResourcePool<HTMLImageElement>
-) => new Promise<Resource<HTMLImageElement>>((resolve) => {
+const loadImage = (url: string, imgPool): Promise<HTMLImageElement> => new Promise((resolve) => {
   const img = imgPool.get()
   img.onload = () => {
     resolve(img)
@@ -60,8 +57,7 @@ const loadImage = (
 })
 
 const getImageBlobFromUrl = async (
-  img: HTMLImageElement, w: number, h: number, fileType: string,
-  canvasPool: ResourcePool<HTMLCanvasElement>
+  img: HTMLImageElement, w: number, h: number, fileType: string, canvasPool
 ): Promise<Blob> => {
   const canvas = canvasPool.get()
   const canvasCtx = canvas.getContext('2d')
