+++ reality/cloud/xrhome/src/client/studio/invalid-image-target-tooltip.tsx
@@ -5,7 +5,7 @@ import type {DeepReadonly} from 'ts-essentials'
 import type {ImageTarget as ImageTargetConfig} from '@ecs/shared/scene-graph'
 
 import {Tooltip} from '../ui/components/tooltip'
-import {useImageTarget} from '../image-targets/use-image-targets'
+import {useImageTarget} from './hooks/use-image-target'
 
 const useStyles = createUseStyles({
   tooltipContent: {
