+++ reality/cloud/xrhome/src/shared/validate-image-target-name.ts
@@ -13,7 +13,7 @@ type ValidateOptions = {
 const validateImageTargetName = (name: string, options?: ValidateOptions) => {
   if (!name) {
     return 'You need to specify name'
-  } else if (options?.otherImageNames?.includes(name)) {
+  } else if (options?.otherImageNames.includes(name)) {
     return 'Another image target has the same name'
   } else if (DISALLOWED_NAME_CHARACTERS.some(char => name.includes(char))) {
     return `Image target name may not contain: ${DISALLOWED_NAME_CHARACTERS.join(', ')}`
